Parser class: extract parser behaviour
Parser functions is handled by main file Aurora.java. These functions
are primarily utility functions that convert data into other data types.

Let's
* Create a new class Parser
* Extract parse lines from storage file
* Extract parse int
* Extract parse date time
